Extends README.md to mention indirect use.
This CL adds a paragraph to README.md explaining that even indirect
clients of reflectable need to specify the reflectable transformer in
their `pubspec.yaml`. This is not necessary for clients who want to run
in pre-transformed mode all the time, but presumably such clients do not
exist and hence we won't help them. In general, indirect usage involves
an intermediate library L, and "specify the reflectable transformer"
should be part of the documentation for L; hence, this CL also adds a
sentence directed at writers of such intermediate libraries.

Adresses issue 84.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/2002033002 .