Adjust usage of InterfaceTypeImpl constructor to match analyzer 0.39.4 and up (#196)
Adjusted 'builder_implementation.dart' to work with new `InterfaceTypeImpl` constructor. Adjusted version number to 2.2.1+1. Updated lower bound on analyzer version to the version that actually has the new constructor signature, setting upper bound to the same version: Reflectable is now pinned to exactly one version of the analyzer, because it needs to use parts of the analyzer that are not considered public (and hence, every change in the analyzer is potentially a breaking change for reflectable).