Generate code with nullsafety (#231)
Add legacy copy of tests in test_reflectable/test in test_reflectable/test/legacy, such that we have tests whose entry point is opted in with null safety as well as tests whose entry point is opted out. Note that the former will cause a hint about reverse migration (some opted-in libraries import some opted-out libraries), but that's unavoidable as long as build_test is opted out.

Note that all tests must be executed with unsound null checking, because none of them can avoid importing some opted-out libraries (so they are all mixed-version programs).

Migrate the opted-in tests to null safety, fixing various compile-time errors.

Migrate code generation and run-time support for null safety. One change in code generation that is not fully forced by the null safety migration is that we now avoid null when an `int` is required; we could have used `int?`, but we're instead using the value -1 (NO_CAPABILITY_INDEX). This means that some situations will appear to be a 'no capability' issue even though they are caused by something else; this should not be observable because those other causes would be associated with values of other reflector data that prevents said entry from being looked up.

As a result, code generation will generate code with null safety or not (when the entry point uses null safety resp. when it does not), but the code generator itself is not yet migrated. The supporting libraries (imported by the programs for which we're generating code) have been migrated.

Update reflectable/example.dart.

Bug fix, unrelated to null safety: Add `.declaration` to go from `...Member` to actual `Element` for `GlobalQuantify...`.