Use nullsafe build, migrate to null safety (#248)
This PR migrates the reflectable code generator to null safety, and it uses the newly arrived null safe versions of build packages, such that reflectable is fully opted in to null safety.

In general, nulls that arise because the program does not pass static analysis are asserted (so there are lots of ways that the code generation will crash with a null error if we run it on a program with errors). We may wish to change this later on, to avoid all those possible crashes. The best approach would probably be to ask the analyzer on startup whether this program has any errors, and only generate code if not.