Fix issue #253 and update dependencies (#254)
Remove a ! and handle the null which may actually occur (fixing #253). Change the analyzer dependency to ^2.0.0, and perform a handful of changes that follow from this.

One test was changed (presumably a bug was fixed in the analyzer): An initializing formal parameter is now reported to be final, and a test expectation has been changed to reflect this new behavior.