Update reflectable to analyzer 3.2.0, cache some getters (#264)
Update reflectable to use version 3.2.0 of the analyzer, and adjust several locations in the code based on the fact that several getters/methods in the analyzer now return a non-null value (so we do not need to check for null any more). Update the version number and CHANGELOG.md such that we can release reflectable 3.0.6 based on this commit.