Update format (#340)
Update package versions such that we can use a recent version of the formatter, in response to https://github.com/google/reflectable.dart/issues/338, and prepare for release as reflectable 4.0.13.

This PR changes reflectable to use language version 3.6. It uses newer versions of `build_runner_core` (^8.0.0) and `dart_style` (^3.0.0), to enable using the new formatter to format the generated code.

This PR changes analysis_options.yaml such that It ignores `deprecated_member_use` because `build_runner` is not yet ready for the new `Element` model. This allows us to keep using the old element model for a while, until it is possible to update to the new model.

This PR also adds support for suppression of warnings via environment variables. The environment variable has prefix `REFLECTABLE_SUPPRESS_...` and then specifies the warning using the name of the warning in `UPPER_CASE_WITH_UNDERSCORES` style. For example, a command like `export REFLECTABLE_SUPPRESS_BAD_METADATA=yes` will suppress the `badMetadata` warning when using `bash` or a similar shell.

It also introduces support for specifying the language version used by the formatter by setting `REFLECTABLE_FORMATTER_LANGUAGE_VERSION` to a value of the form `[0-9]+\\.[0-9]+`. For example, `export REFLECTABLE_FORMATTER_LANGUAGE_VERSION=3.7` will make the formatter use language version 3.7.

Finally, information is added to the design document in `doc` about these environment variables and their effect.