Update reflectable to work with Flutter 3.35 (#343)
This commit changes the version of the analyzer to ^7.0.0, such that there is no dependency on _macros (which doesn't exist with Flutter 3.35, so we really need this step). It makes minimal changes with respect to the new element model (Element2), just enough to make the code generator and generated code work in this new context. This implies that there will be another update where reflectable switches to the new element model (which will allow reflectable to use yet newer versions of the analyser).