Split off the builder (#345)
This commit changes the `reflectable` package such that it only contains run-time related resources. The code generator is moved to a new package `reflectable_builder`. It changes the repository to use a pub workspace. It removes several dependencies that are not used any more. It marks two capabilities as deprecated (they have not been is use for a while). It changes several `dependencies` to `dev_dependencies`, to allow for faster build processes using less memory.

These changes are breaking (in particular, client packages must use the `reflectable_builder` package and the `reflectable_builder` builder, hence the version number 5.0.0 in `reflectable`.

This commit prepares for release of `reflectable` at version 5.0.0 and `reflectable_builder` at version 1.0.0.