Update to analyzer 8 and build 4 (#346)
This PR updates the reflectable_builder code generator to use analyzer version ^8.0.0 and build version ^4.0.0. It compensates for the change whereby getter names omit the `=` at the end of the name. It handles the possible null valued `name` of many elements by using 'unknownName'. It uses `getters` and `setters` in place of the (now removed) `accessors`. Finally, it eliminates the `MixinApplication` wrapper which is not needed any more.